New folder structure and better namescaping
