--- lib/tyrant/contract/mail.rb
@@ -1,16 +1,0 @@
-require 'reform'
-require 'reform/form/dry'
-
-module Tyrant::Contract
-  class Mail < Reform::Form 
-    feature Reform::Form::Dry
-
-    property :email, virtual: true
-    property :new_password, virtual: true
-
-    validation do
-      required(:email).filled
-      required(:new_password).filled
-    end
-  end
-end
