+++ lib/tyrant/mailer.rb
@@ -1,0 +1,2 @@
+require "tyrant/mailer/contract/email"
+require "tyrant/mailer/operation/mailer"
