--- lib/tyrant/operation/get_new_password.rb
@@ -1,6 +1,0 @@
-require 'trailblazer'
-require 'tyrant/contract/change_password'
-
-class Tyrant::GetNewPassword < Trailblazer::Operation
-  step Trailblazer::Operation::Contract::Build(constant: Tyrant::Contract::ChangePassword) 
-end
