+++ lib/tyrant/reset_password.rb
@@ -1,0 +1,2 @@
+require "tyrant/reset_password/contract/get_email"
+require "tyrant/reset_password/operation/reset_password"
