+++ test/authenticatable_test.rb
